chore: move logging builders C bindings into common folder
